+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/locationresource/LocationResourceIntTest.kt
@@ -1,0 +1,1110 @@
+package uk.gov.justice.digital.hmpps.locationsinsideprison.resource.locationresource
+
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.TestConfiguration
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Primary
+import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.PermanentDeactivationLocationRequest
+import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.TemporaryDeactivationLocationRequest
+import uk.gov.justice.digital.hmpps.locationsinsideprison.integration.SqsIntegrationTestBase
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.Capacity
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.Cell
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.Certification
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.DeactivatedReason
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.LocationType
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.NonResidentialLocation
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.NonResidentialUsageType
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.ResidentialAttributeValue
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.ResidentialLocation
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.SpecialistCellType
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.LocationRepository
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.buildCell
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.buildNonResidentialLocation
+import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.buildResidentialLocation
+import uk.gov.justice.hmpps.test.kotlin.auth.WithMockAuthUser
+import java.time.Clock
+import java.time.LocalDate
+import java.time.LocalDateTime
+
+const val EXPECTED_USERNAME = "A_TEST_USER"
+
+@WithMockAuthUser(username = EXPECTED_USERNAME)
+class LocationResourceIntTest : SqsIntegrationTestBase() {
+
+  @TestConfiguration
+  class FixedClockConfig {
+    @Primary
+    @Bean
+    fun fixedClock(): Clock = clock
+  }
+
+  @Autowired
+  lateinit var repository: LocationRepository
+  lateinit var cell1: Cell
+  lateinit var cell2: Cell
+  lateinit var inactiveCellB3001: Cell
+  lateinit var archivedCell: Cell
+  lateinit var landingZ1: ResidentialLocation
+  lateinit var landingZ2: ResidentialLocation
+  lateinit var landingB3: ResidentialLocation
+  lateinit var landingN1: ResidentialLocation
+  lateinit var wingZ: ResidentialLocation
+  lateinit var wingB: ResidentialLocation
+  lateinit var wingN: ResidentialLocation
+  lateinit var visitRoom: NonResidentialLocation
+  lateinit var adjRoom: NonResidentialLocation
+
+  @BeforeEach
+  fun setUp() {
+    prisonerSearchMockServer.resetAll()
+    repository.deleteAll()
+
+    wingN = repository.save(
+      buildResidentialLocation(
+        prisonId = "NMI",
+        pathHierarchy = "A",
+        locationType = LocationType.WING,
+        localName = "WING A",
+      ),
+    )
+
+    landingN1 = repository.save(
+      buildResidentialLocation(
+        prisonId = "NMI",
+        pathHierarchy = "A-1",
+        locationType = LocationType.LANDING,
+        localName = "LANDING A",
+      ),
+    )
+
+    wingN.addChildLocation(landingN1)
+    repository.save(wingN)
+
+    wingZ = repository.save(
+      buildResidentialLocation(
+        pathHierarchy = "Z",
+        locationType = LocationType.WING,
+      ),
+    )
+    wingB = repository.save(
+      buildResidentialLocation(
+        pathHierarchy = "B",
+        locationType = LocationType.WING,
+      ),
+    )
+    landingZ1 = repository.save(
+      buildResidentialLocation(
+        pathHierarchy = "Z-1",
+        locationType = LocationType.LANDING,
+      ),
+    )
+    landingZ2 = repository.save(
+      buildResidentialLocation(
+        pathHierarchy = "Z-2",
+        locationType = LocationType.LANDING,
+      ),
+    )
+    landingB3 = repository.save(
+      buildResidentialLocation(
+        pathHierarchy = "B-A",
+        locationType = LocationType.LANDING,
+      ),
+    )
+    cell1 = repository.save(
+      buildCell(
+        pathHierarchy = "Z-1-001",
+        capacity = Capacity(maxCapacity = 2, workingCapacity = 2),
+        certification = Certification(certified = true, capacityOfCertifiedCell = 2),
+      ),
+    )
+    cell2 = repository.save(
+      buildCell(
+        pathHierarchy = "Z-1-002",
+        capacity = Capacity(maxCapacity = 2, workingCapacity = 2),
+        certification = Certification(certified = true, capacityOfCertifiedCell = 2),
+        residentialAttributeValues = setOf(
+          ResidentialAttributeValue.CAT_A,
+          ResidentialAttributeValue.SAFE_CELL,
+          ResidentialAttributeValue.DOUBLE_OCCUPANCY,
+        ),
+        specialistCellType = SpecialistCellType.ACCESSIBLE_CELL,
+      ),
+    )
+    inactiveCellB3001 = repository.save(
+      buildCell(
+        pathHierarchy = "B-A-001",
+        active = false,
+        capacity = Capacity(maxCapacity = 2, workingCapacity = 2),
+        certification = Certification(certified = true, capacityOfCertifiedCell = 2),
+        specialistCellType = SpecialistCellType.ACCESSIBLE_CELL,
+      ),
+    )
+
+    archivedCell = repository.save(
+      buildCell(
+        pathHierarchy = "Z-1-003",
+        capacity = Capacity(maxCapacity = 2, workingCapacity = 2),
+        certification = Certification(certified = true, capacityOfCertifiedCell = 2),
+        active = false,
+        archived = true,
+      ),
+    )
+
+    visitRoom = repository.save(
+      buildNonResidentialLocation(
+        pathHierarchy = "VISIT",
+        locationType = LocationType.VISITS,
+        nonResidentialUsageType = NonResidentialUsageType.VISIT,
+      ),
+    )
+    adjRoom = repository.save(
+      buildNonResidentialLocation(
+        pathHierarchy = "ADJUDICATION",
+        locationType = LocationType.ADJUDICATION_ROOM,
+        nonResidentialUsageType = NonResidentialUsageType.ADJUDICATION_HEARING,
+      ),
+    )
+    wingZ.addChildLocation(visitRoom)
+      .addChildLocation(
+        landingZ1
+          .addChildLocation(cell1)
+          .addChildLocation(cell2),
+      )
+      .addChildLocation(landingZ2)
+
+    wingZ.updateComments(
+      "A New Comment",
+      uk.gov.justice.digital.hmpps.locationsinsideprison.resource.locationresource.EXPECTED_USERNAME,
+      clock,
+    )
+
+    wingB.addChildLocation(landingB3.addChildLocation(inactiveCellB3001))
+    repository.save(wingZ)
+    repository.save(wingB)
+  }
+
+  @DisplayName("GET /locations")
+  @Nested
+  inner class ViewPagedLocationsTest {
+
+    @Nested
+    inner class Security {
+
+      @Test
+      fun `access forbidden when no authority`() {
+        webTestClient.get().uri("/locations")
+          .exchange()
+          .expectStatus().isUnauthorized
+      }
+
+      @Test
+      fun `access forbidden when no role`() {
+        webTestClient.get().uri("/locations")
+          .headers(setAuthorisation(roles = listOf()))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden with wrong role`() {
+        webTestClient.get().uri("/locations")
+          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS")))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+    }
+
+    @Nested
+    inner class HappyPath {
+
+      @Test
+      fun `can retrieve details of a page of locations`() {
+        webTestClient.get().uri("/locations?size=13&sort=pathHierarchy,asc")
+          .headers(setAuthorisation(roles = listOf("ROLE_VIEW_LOCATIONS")))
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+              {
+                "totalPages": 1,
+                "totalElements": 13,
+                "first": true,
+                "last": true,
+                "size": 13,
+                "content": [
+                  {
+                    "prisonId": "NMI",
+                    "code": "A",
+                    "pathHierarchy": "A",
+                    "locationType": "WING",
+                    "localName": "WING A",
+                    "key": "NMI-A"
+                  }, 
+                  {
+                    "prisonId": "NMI",
+                    "code": "1",
+                    "pathHierarchy": "A-1",
+                    "locationType": "LANDING",
+                    "localName": "LANDING A",
+                    "key": "NMI-A-1"
+                  }, 
+                  {
+                    "prisonId": "MDI",
+                    "code": "ADJUDICATION",
+                    "pathHierarchy": "ADJUDICATION",
+                    "locationType": "ADJUDICATION_ROOM",
+                    "key": "MDI-ADJUDICATION"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "Z",
+                    "pathHierarchy": "Z",
+                    "locationType": "WING",
+                    "key": "MDI-Z"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "B",
+                    "pathHierarchy": "B",
+                    "locationType": "WING",
+                    "key": "MDI-B"
+                  },
+                  { 
+                    "prisonId": "MDI",
+                    "code": "VISIT",
+                    "pathHierarchy": "Z-VISIT",
+                    "locationType": "VISITS",
+                    "key": "MDI-Z-VISIT"
+                  }, 
+                  {
+                    "prisonId": "MDI",
+                    "code": "1",
+                    "pathHierarchy": "Z-1",
+                    "locationType": "LANDING",
+                    "key": "MDI-Z-1"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "A",
+                    "pathHierarchy": "B-A",
+                    "locationType": "LANDING",
+                    "key": "MDI-B-A"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "2",
+                    "pathHierarchy": "Z-2",
+                    "locationType": "LANDING",
+                    "key": "MDI-Z-2"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "001",
+                    "pathHierarchy": "Z-1-001",
+                    "locationType": "CELL",
+                    "key": "MDI-Z-1-001"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "002",
+                    "pathHierarchy": "Z-1-002",
+                    "locationType": "CELL",
+                    "key": "MDI-Z-1-002"
+                  }, 
+                  {
+                    "prisonId": "MDI",
+                    "code": "003",
+                    "pathHierarchy": "Z-1-003",
+                    "locationType": "CELL",
+                    "key": "MDI-Z-1-003"
+                  },
+                  {
+                    "prisonId": "MDI",
+                    "code": "001",
+                    "pathHierarchy": "B-A-001",
+                    "locationType": "CELL",
+                    "key": "MDI-B-A-001"
+                  }
+                ],
+                "number": 0,
+                "sort": {
+                  "empty": false,
+                  "sorted": true,
+                  "unsorted": false
+                },
+                "numberOfElements": 13,
+                "pageable": {
+                  "offset": 0,
+                  "sort": {
+                    "empty": false,
+                    "sorted": true,
+                    "unsorted": false
+                  },
+                  "pageSize": 13,
+                  "pageNumber": 0,
+                  "paged": true,
+                  "unpaged": false
+                },
+                "empty": false
+              }
+
+              """,
+            false,
+          )
+      }
+    }
+  }
+
+  @DisplayName("GET /locations/{id}")
+  @Nested
+  inner class ViewLocationTest {
+
+    @Nested
+    inner class Security {
+
+      @Test
+      fun `access forbidden when no authority`() {
+        webTestClient.get().uri("/locations/${cell1.id}")
+          .exchange()
+          .expectStatus().isUnauthorized
+      }
+
+      @Test
+      fun `access forbidden when no role`() {
+        webTestClient.get().uri("/locations/${cell1.id}")
+          .headers(setAuthorisation(roles = listOf()))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden with wrong role`() {
+        webTestClient.get().uri("/locations/${cell1.id}")
+          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS")))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+    }
+
+    @Nested
+    inner class HappyPath {
+
+      @Test
+      fun `can retrieve details of a location`() {
+        webTestClient.get().uri("/locations/${wingZ.id}?includeChildren=true")
+          .headers(setAuthorisation(roles = listOf("ROLE_VIEW_LOCATIONS")))
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+             {
+              "prisonId": "MDI",
+              "code": "Z",
+              "pathHierarchy": "Z",
+              "locationType": "WING",
+              "active": true,
+              "key": "MDI-Z",
+              "capacity": {
+                "maxCapacity": 4,
+                "workingCapacity": 4
+              },
+              "certification": {
+                "capacityOfCertifiedCell": 4
+              },
+              "accommodationTypes": [
+                "NORMAL_ACCOMMODATION"
+              ],
+              "usedFor": [
+                "STANDARD_ACCOMMODATION"
+              ],
+              "specialistCellTypes": [
+                "ACCESSIBLE_CELL"
+              ],
+              "childLocations": [
+                {
+                  "prisonId": "MDI",
+                  "code": "VISIT",
+                  "pathHierarchy": "Z-VISIT",
+                  "locationType": "VISITS",
+                  
+                  "active": true,
+                  "key": "MDI-Z-VISIT"
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "1",
+                  "pathHierarchy": "Z-1",
+                  "locationType": "LANDING",
+                  "active": true,
+                  "key": "MDI-Z-1",
+                  "capacity": {
+                    "maxCapacity": 4,
+                    "workingCapacity": 4
+                  },
+                  "certification": {
+                    "capacityOfCertifiedCell": 4
+                  },
+                  "childLocations": [
+                    {
+                      "prisonId": "MDI",
+                      "code": "001",
+                      "pathHierarchy": "Z-1-001",
+                      "locationType": "CELL",
+                      "active": true,
+                      "key": "MDI-Z-1-001",
+                      "capacity": {
+                        "maxCapacity": 2,
+                        "workingCapacity": 2
+                      },
+                      "certification": {
+                        "certified": true,
+                        "capacityOfCertifiedCell": 2
+                      }
+                    },
+                    {
+                      "prisonId": "MDI",
+                      "code": "002",
+                      "pathHierarchy": "Z-1-002",
+                      "locationType": "CELL",
+                      "active": true,
+                      "key": "MDI-Z-1-002",
+                      "capacity": {
+                        "maxCapacity": 2,
+                        "workingCapacity": 2
+                      },
+                      "certification": {
+                        "certified": true,
+                        "capacityOfCertifiedCell": 2
+                      }
+                    }
+                  ]
+                },
+                { 
+                  "prisonId": "MDI",
+                  "code": "2",
+                  "pathHierarchy": "Z-2",
+                  "locationType": "LANDING",
+                  "accommodationTypes":[],
+                  "capacity": {
+                    "maxCapacity": 0,
+                    "workingCapacity": 0
+                  },
+                  "certification": {
+                    "certified": false,
+                    "capacityOfCertifiedCell": 0
+                  },
+                  
+                  "active": true,
+                  "childLocations": [],
+                  "isResidential": true,
+                  "key": "MDI-Z-2"
+                }
+              ]
+            }
+          """,
+            false,
+          )
+      }
+    }
+  }
+
+  @DisplayName("PUT /locations/{id}/deactivate")
+  @Nested
+  inner class TemporarilyDeactivateLocationTest {
+
+    @Nested
+    inner class Security {
+      @Test
+      fun `access forbidden when no authority`() {
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .exchange()
+          .expectStatus().isUnauthorized
+      }
+
+      @Test
+      fun `access forbidden when no role`() {
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf()))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.DAMAGED)))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden with wrong role`() {
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.DAMAGED)))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden with right role, wrong scope`() {
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS"), scopes = listOf("read")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.DAMAGED)))
+          .exchange()
+          .expectStatus().isForbidden
+      }
+    }
+
+    @Nested
+    inner class Validation {
+      @Test
+      fun `access client error bad data`() {
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue("""{"deactivationReason": ""}""")
+          .exchange()
+          .expectStatus().is4xxClientError
+      }
+
+      @Test
+      fun `cannot deactivate a location when prisoner is inside the cell`() {
+        prisonerSearchMockServer.stubSearchByLocations(cell1.prisonId, listOf(cell1.getPathHierarchy()), true)
+
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/permanent")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(PermanentDeactivationLocationRequest(reason = "Demolished")))
+          .exchange()
+          .expectStatus().isEqualTo(409)
+      }
+
+      @Test
+      fun `cannot deactivate a wing when prisoners are in cells below`() {
+        prisonerSearchMockServer.stubSearchByLocations(wingZ.prisonId, listOf(cell1.getPathHierarchy(), cell2.getPathHierarchy()), true)
+
+        webTestClient.put().uri("/locations/${wingZ.id}/deactivate/permanent")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(PermanentDeactivationLocationRequest(reason = "Demolished")))
+          .exchange()
+          .expectStatus().isEqualTo(409)
+      }
+    }
+
+    @Nested
+    inner class HappyPath {
+      @Test
+      fun `can deactivate a location`() {
+        prisonerSearchMockServer.stubSearchByLocations(cell1.prisonId, listOf(cell1.getPathHierarchy()), false)
+
+        val now = LocalDateTime.now(clock)
+        val proposedReactivationDate = now.plusMonths(1).toLocalDate()
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/permanent")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(PermanentDeactivationLocationRequest(reason = "Cell destroyed")))
+          .exchange()
+          .expectStatus().isOk
+
+        prisonerSearchMockServer.resetAll()
+        prisonerSearchMockServer.stubSearchByLocations(wingZ.prisonId, listOf(cell1.getPathHierarchy(), cell2.getPathHierarchy()), false)
+
+        webTestClient.put().uri("/locations/${wingZ.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.DAMAGED, proposedReactivationDate = proposedReactivationDate)))
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+              {
+                "id": "${wingZ.id}",
+                "prisonId": "MDI",
+                "code": "Z",
+                "pathHierarchy": "Z",
+                "locationType": "WING",
+                "accommodationTypes": [ "NORMAL_ACCOMMODATION" ],
+                "capacity": {
+                  "maxCapacity": 2,
+                  "workingCapacity": 0
+                },
+                "certification": {
+                  "certified": true,
+                  "capacityOfCertifiedCell": 2
+                },
+                "active": false,
+                "deactivatedDate": "$now",
+                "deactivatedReason": "DAMAGED",
+                "permanentlyInactive": false,
+                "proposedReactivationDate": "$proposedReactivationDate",
+                "topLevelId": "${wingZ.id}",
+                "isResidential": true,
+                "key": "MDI-Z",
+                "deactivatedBy": "LOCATIONS_INSIDE_PRISON_API"
+              }
+          """,
+            false,
+          )
+
+        getDomainEvents(8).let {
+          assertThat(it.map { message -> message.eventType to message.additionalInformation?.key }).containsExactlyInAnyOrder(
+            "location.inside.prison.deactivated" to "MDI-Z",
+            "location.inside.prison.deactivated" to "MDI-Z-1",
+            "location.inside.prison.deactivated" to "MDI-Z-2",
+            "location.inside.prison.deactivated" to "MDI-Z-1-001",
+            "location.inside.prison.deactivated" to "MDI-Z-1-002",
+            "location.inside.prison.deactivated" to "MDI-Z-VISIT",
+            "location.inside.prison.amended" to "MDI-Z-1",
+            "location.inside.prison.amended" to "MDI-Z",
+          )
+        }
+
+        webTestClient.get().uri("/locations/${wingZ.id}?includeChildren=true")
+          .headers(setAuthorisation(roles = listOf("ROLE_VIEW_LOCATIONS")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+             {
+              "prisonId": "MDI",
+              "code": "Z",
+              "pathHierarchy": "Z",
+              "locationType": "WING",
+              "active": false,
+              "deactivatedByParent": false,
+              "key": "MDI-Z",
+              "deactivatedReason": "DAMAGED",
+              "accommodationTypes":["NORMAL_ACCOMMODATION"],
+              "permanentlyInactive": false,
+              "proposedReactivationDate": "$proposedReactivationDate",
+              "deactivatedDate": "$now",
+              "capacity": {
+                "maxCapacity": 2,
+                "workingCapacity": 0
+              },
+              "certification": {
+                "certified": true,
+                "capacityOfCertifiedCell": 2
+              },
+              "childLocations": [
+                {
+                  "prisonId": "MDI",
+                  "code": "VISIT",
+                  "pathHierarchy": "Z-VISIT",
+                  "locationType": "VISITS",
+                  "active": false,
+                  "deactivatedByParent": true,
+                  "proposedReactivationDate": "$proposedReactivationDate",
+                  "deactivatedDate": "$now",
+                  "deactivatedReason": "DAMAGED",
+                  "permanentlyInactive": false,
+                  "isResidential": false,
+                  "key": "MDI-Z-VISIT"
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "1",
+                  "pathHierarchy": "Z-1",
+                  "locationType": "LANDING",
+                  "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                  "active": false,
+                  "deactivatedByParent": false,
+                  "proposedReactivationDate": "$proposedReactivationDate",
+                  "deactivatedDate": "$now",
+                  "deactivatedReason": "DAMAGED",
+                  "permanentlyInactive": false,
+                  "isResidential": true,
+                  "key": "MDI-Z-1",
+                  "childLocations": [
+                    {
+                      "prisonId": "MDI",
+                      "code": "002",
+                      "pathHierarchy": "Z-1-002",
+                      "locationType": "CELL",
+                      "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                      "active": false,
+                      "deactivatedByParent": false,
+                      "proposedReactivationDate": "$proposedReactivationDate",
+                      "deactivatedDate": "$now",
+                      "deactivatedReason": "DAMAGED",
+                      "isResidential": true,
+                      "key": "MDI-Z-1-002"
+                    }
+                  ]
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "2",
+                  "pathHierarchy": "Z-2",
+                  "locationType": "LANDING",
+                  "accommodationTypes":[],
+                  "active": false,
+                  "deactivatedByParent": false,
+                  "proposedReactivationDate": "$proposedReactivationDate",
+                  "deactivatedDate": "$now",
+                  "deactivatedReason": "DAMAGED",
+                  "isResidential": true,
+                  "key": "MDI-Z-2"
+                }
+              ]
+            }
+          """,
+            false,
+          )
+      }
+
+      @Test
+      fun `can update a deactivated location`() {
+        prisonerSearchMockServer.stubSearchByLocations(cell1.prisonId, listOf(cell1.getPathHierarchy()), false)
+
+        val now = LocalDateTime.now(clock)
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.DAMAGED)))
+          .exchange()
+          .expectStatus().isOk
+
+        val proposedReactivationDate = now.plusMonths(1).toLocalDate()
+        webTestClient.put().uri("/locations/${cell1.id}/update/temporary-deactivation")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.MOTHBALLED, planetFmReference = "334423", proposedReactivationDate = proposedReactivationDate)))
+          .exchange()
+          .expectStatus().isOk
+
+        getDomainEvents(4).let {
+          assertThat(it.map { message -> message.eventType to message.additionalInformation?.key }).containsExactlyInAnyOrder(
+            "location.inside.prison.amended" to "MDI-Z-1",
+            "location.inside.prison.amended" to "MDI-Z",
+            "location.inside.prison.deactivated" to "MDI-Z-1-001",
+            "location.inside.prison.amended" to "MDI-Z-1-001",
+          )
+        }
+
+        webTestClient.get().uri("/locations/${cell1.id}?includeChildren=true")
+          .headers(setAuthorisation(roles = listOf("ROLE_VIEW_LOCATIONS")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+             {
+                "prisonId": "MDI",
+                "code": "001",
+                "pathHierarchy": "Z-1-001",
+                "locationType": "CELL",
+                "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                "active": false,
+                "deactivatedByParent": false,
+                "proposedReactivationDate": "$proposedReactivationDate",
+                "deactivatedDate": "$now",
+                "deactivatedReason": "MOTHBALLED",
+                "planetFmReference": "334423",
+                "isResidential": true,
+                "key": "MDI-Z-1-001"
+            }
+          """,
+            false,
+          )
+      }
+    }
+  }
+
+  @DisplayName("PUT /locations/{id}/reactivate")
+  @Nested
+  inner class ReactivateLocationTest {
+
+    @Nested
+    inner class Security {
+      @Test
+      fun `access forbidden when no authority`() {
+        webTestClient.put().uri("/locations/${cell1.id}/reactivate")
+          .exchange()
+          .expectStatus().isUnauthorized
+      }
+
+      @Test
+      fun `access forbidden when no role`() {
+        webTestClient.put().uri("/locations/${cell1.id}/reactivate")
+          .headers(setAuthorisation(roles = listOf()))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden with wrong role`() {
+        webTestClient.put().uri("/locations/${cell1.id}/reactivate")
+          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden with right role, wrong scope`() {
+        webTestClient.put().uri("/locations/${cell1.id}/reactivate")
+          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS"), scopes = listOf("read")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isForbidden
+      }
+    }
+
+    @Nested
+    inner class HappyPath {
+      @Test
+      fun `can cascade reactivated locations`() {
+        prisonerSearchMockServer.stubSearchByLocations(cell1.prisonId, listOf(cell1.getPathHierarchy(), cell2.getPathHierarchy()), false)
+
+        webTestClient.put().uri("/locations/${wingZ.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.MOTHBALLED)))
+          .exchange()
+          .expectStatus().isOk
+
+        webTestClient.put().uri("/locations/${wingZ.id}/reactivate?cascade-reactivation=true")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isOk
+
+        getDomainEvents(12).let {
+          assertThat(it.map { message -> message.eventType to message.additionalInformation?.key }).containsExactlyInAnyOrder(
+            "location.inside.prison.deactivated" to "MDI-Z",
+            "location.inside.prison.deactivated" to "MDI-Z-1",
+            "location.inside.prison.deactivated" to "MDI-Z-2",
+            "location.inside.prison.deactivated" to "MDI-Z-1-001",
+            "location.inside.prison.deactivated" to "MDI-Z-1-002",
+            "location.inside.prison.deactivated" to "MDI-Z-VISIT",
+            "location.inside.prison.reactivated" to "MDI-Z",
+            "location.inside.prison.reactivated" to "MDI-Z-1",
+            "location.inside.prison.reactivated" to "MDI-Z-2",
+            "location.inside.prison.reactivated" to "MDI-Z-1-001",
+            "location.inside.prison.reactivated" to "MDI-Z-1-002",
+            "location.inside.prison.reactivated" to "MDI-Z-VISIT",
+          )
+        }
+
+        webTestClient.get().uri("/locations/${wingZ.id}?includeChildren=true")
+          .headers(setAuthorisation(roles = listOf("ROLE_VIEW_LOCATIONS")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+             {
+              "prisonId": "MDI",
+              "code": "Z",
+              "pathHierarchy": "Z",
+              "locationType": "WING",
+              "active": true,
+              "key": "MDI-Z",
+              "capacity": {
+                "maxCapacity": 4,
+                "workingCapacity": 4
+              },
+              "certification": {
+                "capacityOfCertifiedCell": 4
+              },
+              "childLocations": [
+                {
+                  "prisonId": "MDI",
+                  "code": "VISIT",
+                  "pathHierarchy": "Z-VISIT",
+                  "locationType": "VISITS",
+                  "active": true,
+                  "isResidential": false,
+                  "key": "MDI-Z-VISIT"
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "1",
+                  "pathHierarchy": "Z-1",
+                  "locationType": "LANDING",
+                  "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                  "active": true,
+                  "isResidential": true,
+                  "key": "MDI-Z-1",
+                  "childLocations": [
+                    {
+                      "prisonId": "MDI",
+                      "code": "001",
+                      "pathHierarchy": "Z-1-001",
+                      "locationType": "CELL",
+                      "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                      "active": true,
+                      "isResidential": true,
+                      "key": "MDI-Z-1-001"
+                    },
+                    {
+                      "prisonId": "MDI",
+                      "code": "002",
+                      "pathHierarchy": "Z-1-002",
+                      "locationType": "CELL",
+                      "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                      "active": true,
+                      "isResidential": true,
+                      "key": "MDI-Z-1-002"
+                    }
+                  ]
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "2",
+                  "pathHierarchy": "Z-2",
+                  "locationType": "LANDING",
+                  "accommodationTypes":[],
+                  "active": true,
+                  "isResidential": true,
+                  "key": "MDI-Z-2"
+                }
+              ]
+            }
+          """,
+            false,
+          )
+      }
+
+      @Test
+      fun `can reactivate a location`() {
+        prisonerSearchMockServer.stubSearchByLocations(cell1.prisonId, listOf(cell1.getPathHierarchy()), false)
+
+        val proposedReactivationDate = LocalDate.now(clock).plusMonths(1)
+        webTestClient.put().uri("/locations/${cell1.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.DAMAGED, proposedReactivationDate = proposedReactivationDate)))
+          .exchange()
+          .expectStatus().isOk
+
+        prisonerSearchMockServer.resetAll()
+        prisonerSearchMockServer.stubSearchByLocations(wingZ.prisonId, listOf(cell2.getPathHierarchy(), cell1.getPathHierarchy()), false)
+
+        webTestClient.put().uri("/locations/${wingZ.id}/deactivate/temporary")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .bodyValue(jsonString(TemporaryDeactivationLocationRequest(deactivationReason = DeactivatedReason.MOTHBALLED)))
+          .exchange()
+          .expectStatus().isOk
+
+        webTestClient.put().uri("/locations/${cell1.id}/reactivate")
+          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isOk
+
+        getDomainEvents(12).let {
+          assertThat(it.map { message -> message.eventType to message.additionalInformation?.key }).containsExactlyInAnyOrder(
+            "location.inside.prison.deactivated" to "MDI-Z-1-001",
+            "location.inside.prison.amended" to "MDI-Z-1",
+            "location.inside.prison.amended" to "MDI-Z",
+            "location.inside.prison.deactivated" to "MDI-Z",
+            "location.inside.prison.deactivated" to "MDI-Z-1",
+            "location.inside.prison.deactivated" to "MDI-Z-2",
+            "location.inside.prison.deactivated" to "MDI-Z-1-001",
+            "location.inside.prison.deactivated" to "MDI-Z-1-002",
+            "location.inside.prison.deactivated" to "MDI-Z-VISIT",
+            "location.inside.prison.reactivated" to "MDI-Z-1-001",
+            "location.inside.prison.amended" to "MDI-Z-1",
+            "location.inside.prison.amended" to "MDI-Z",
+          )
+        }
+
+        webTestClient.get().uri("/locations/${wingZ.id}?includeChildren=true")
+          .headers(setAuthorisation(roles = listOf("ROLE_VIEW_LOCATIONS")))
+          .header("Content-Type", "application/json")
+          .exchange()
+          .expectStatus().isOk
+          .expectBody().json(
+            // language=json
+            """
+             {
+              "prisonId": "MDI",
+              "code": "Z",
+              "pathHierarchy": "Z",
+              "locationType": "WING",
+              "active": true,
+              "key": "MDI-Z",
+              "capacity": {
+                "maxCapacity": 4,
+                "workingCapacity": 2
+              },
+              "certification": {
+                "capacityOfCertifiedCell": 4
+              },
+              "childLocations": [
+                {
+                  "prisonId": "MDI",
+                  "code": "VISIT",
+                  "pathHierarchy": "Z-VISIT",
+                  "locationType": "VISITS",
+                  "active": true,
+                  "isResidential": false,
+                  "key": "MDI-Z-VISIT"
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "1",
+                  "pathHierarchy": "Z-1",
+                  "locationType": "LANDING",
+                  "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                  "active": true,
+                  "isResidential": true,
+                  "key": "MDI-Z-1",
+                  "childLocations": [
+                    {
+                      "prisonId": "MDI",
+                      "code": "001",
+                      "pathHierarchy": "Z-1-001",
+                      "locationType": "CELL",
+                      "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                      "active": true,
+                      "isResidential": true,
+                      "key": "MDI-Z-1-001"
+                    },
+                    {
+                      "prisonId": "MDI",
+                      "code": "002",
+                      "pathHierarchy": "Z-1-002",
+                      "locationType": "CELL",
+                      "accommodationTypes":["NORMAL_ACCOMMODATION"],
+                      "active": false,
+                      "deactivatedReason": "MOTHBALLED",
+                      "isResidential": true,
+                      "key": "MDI-Z-1-002"
+                    }
+                  ]
+                },
+                {
+                  "prisonId": "MDI",
+                  "code": "2",
+                  "pathHierarchy": "Z-2",
+                  "locationType": "LANDING",
+                  "accommodationTypes":[],
+                  "active": false,
+                  "deactivatedReason": "MOTHBALLED",
+                  "isResidential": true,
+                  "key": "MDI-Z-2"
+                }
+              ]
+            }
+          """,
+            false,
+          )
+      }
+    }
+  }
+}
